fix: restore just ci after refactor/compiler-update merge
Tests and clippy broke after the merge folded in uor-foundation 0.1.4,
the new HoloHeader fields (certificate_offset / certificate_size /
unit_address), and the cascade-era enum/helper changes. This commit
drags all of that back to a green state.

Tests:
- hologram-ring/tests/uor_conformance.rs: the trait `Datum::spectrum`
  now returns `P::NonNegativeInteger` (u64) — assertion updated, and
  the inherent `Datum::spectrum()` &str form is verified alongside.
  `AlgebraCommutator` / `AlgebraAssociator` are marker traits in 0.1.4,
  so the former `*_formula()` tests became trait-impl smoke tests.
- hologram-archive/tests/blake3_quantum.rs and
  hologram-exec/tests/hr_contract.rs: added `certificate_offset: 0`,
  `certificate_size: 0`, and `unit_address: [0u8; 32]` to every
  `HoloHeader` struct literal (8 sites total).
- hologram-exec/src/kv/store.rs tests and hologram-cascade/src/qedl.rs
  test: replaced the now-removed `GraphOp::MatMulLut { bits, cid }` with
  the per-precision `MatMulLut4` / `MatMulLut8` variants.

Clippy (was firing under `-D warnings`):
- hologram-core: declared `no_alloc` as a feature so the cfg-check stops
  flagging it across term/arena.rs (29 warnings-as-errors cleared).
- hologram-core/src/term/enforcement_bridge.rs: dropped a `u32 -> u32`
  no-op cast.
- hologram-graph/src/fusion/cse.rs + schedule/levels.rs: removed
  needless `&` on an already-`&[…]` slice.
- hologram-cascade/src/certificate.rs: `for slot in &self.slots { if let
  Some(...) }` → `for (key, cert) in self.slots.iter().flatten()`.
- hologram-cascade/src/dispatch_decl.rs + effect_decl.rs: added
  `impl Default` wrapping the existing `new()` for `DispatchRegistry`
  and `EffectStore`.
- hologram-compiler/src/compiler/mod.rs: `#[allow(clippy::large_enum_
  variant)]` on `CompilerInput`, since the `Unit` variant holds a ~2 KB
  `HoloCompileUnit` and boxing would just add indirection for a
  builder-only payload.

`just ci` now passes: fmt-check + clippy (no denied warnings) + 1674
tests, 0 failures.

+++ crates/hologram-core/src/term/enforcement_bridge.rs
@@ -134,7 +134,7 @@ pub fn lower_enforcement_node<const CAP: usize>(
                 // Last body is the default accumulator
                 let last_body_idx = arms.start + arms.len - 1;
                 let mut acc = lower_enforcement_node(last_body_idx, src, dst, node_map)?;
-                let num_arms = (arms.len / 2) as u32;
+                let num_arms = arms.len / 2;
                 // Process arms right-to-left: each arm selects body when
                 // scrutinee == pattern via XOR-to-zero + mask
                 for i in (0..num_arms).rev() {
